// src/types/TpleType.enum.ts
export enum TplType {
  /** SearchPage过滤器 */
  TYPE_FILTER_FORM = "FilterForm",
  /** SearchPage内容部分 */
  TYPE_CONTENT = "Content",
  /** SearchPage入口 */
  TYPE_ENTRANCE = "Entrance",
  /** renderConf配置[仅在使用某些选项时起作用] */
  TYPE_RENDER_CONF = "renderConf",
}